requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.16.1

# Flask Packages
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Flask-User==1.0.1.5
Flask-Bootstrap4

# SQLAlchemy
SQLAlchemy==1.4.17
PyMySQL==1.0.2

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0